Fixed bug where if you shift clicked a feature when nothing else was selected, that feature would appear selected but in fact it wasnt, and it wouldnt deselect. This was particularly annoying on doing multiselects where your instinct is to shift click the 1st item even though you dont have to. There was 2 things that caused this problem. When nothing else is selected both mousePressed and mouseClicked would call doLeftClick. Basically you would get 2 selections for one click. I took the doLeftClick out of mousePressed. This is ok for dragging because as soon as it starts dragging doViewDrag calls mouseClicked which does the selection. In the past there was an inconsistency in that if you pressed the mouse (without release) and there was nothing else selected it would select, where if there was other stuff selected it wouldnt select.
The second part of this problem was the force issue in
Selection. Selection.add has a force boolean, that if false will cause
a SelectionItem to be removed if it is already present. In other words
2 adds make a remove. This with the above issue made the bug. So
SelectionManager should have been calling Selection with
force=true. In fact nowhere is force called with false, except this
mistake in SelectionManager. So I took out the force boolean, so its
always force=true, for clarity sake. So if something is added that is
already selected, it just doesnt add it, but it doesnt remove
it. Deselection happens separately, not through the add method.