Finished up work that Nomi had started on getting move to opposite stand working again for results. I added functions to FeatureSetI to record a feature sets analogous feature set on the opposite strand. This facilitates getting the new parent on the opposite strand. (has/get/setAnalogousOppositeStrandFeatureset)
Since DrawableFeatureSet implements FeatureSetI I had to add the
methods there as well. It might be nice to have DrawableFeatureSet not
implement FeatureSetI. Presently, it blurs the model and view. When you have a
FeatureSetI you dont know if you actually have model. You have to test
for instanceof Drawable, and if so cast to the drawable and ask for
its model. The drawable could still wrap its models functions for
convenience but it might be more clear not to implement the model's
interface.