The highlight line for the ede in szap was off by one base. This is now fixed. It turns out it was a transformer issue. toPixelX (or toPixel.x) returns the rightmost/max pixel value for the user coordinate/basepair. This is problematic for the beginning of a range that wants to be inclusive as in this case. You have to know to subtract one to get it right, which is awkward.
The minXPixelAtUserCoord method in Transformer was clearly meant to
address this issue, though it had not been implemented yet(it just
returned -1). I put in
an implementation that gives the rightmost/minimum pixel value for the
bp, and szap.calculateHighlightRect now uses that, rather than
subtracting one. In fact all that minXPixel does is call
toPixelX(user-1), but it makes the issue clearer.

One could imagine going further with making this clearer by having
methods that take in stranded user ranges and return pixel (stranded?)
ranges, or something like that, as a convenience function.