Fixed null pointer on deleting of selection that was shift double clicked. The problem was the selection included both a transcript and exon. Having both parent and child messed up the editors deleting function. So Selection now checks if the addition to the selection is contained or contains anything already selected. I think this might fix a lot of other amorphous non-repeatable editing bugs.
I reworked FeatureSet's findFeature function for this cause. It was
not being used anywhere. So I renamed it "contains(sf)" and had it return a boolean.
I also added it to SeqFeature and SeqFeatureI. For SeqFeature it just
test equality. FeatureSet overrides this and recurses through
descendants as well as equality.