Fixed bug where annot menu items were disabled on external gene selection (like gene menu selection). The problem was AnnotMenu was using Selection.getSelectionForSource(annotView).
getSelectionForSource is problematic. Its being used to get all things
selected in that view. But just because something is selected in that
does not at all mean that that view is the source of the
selection. This is a fundamental apollo flaw at the moment. The
annotation editor relies heavily on this faulty assumption, and thus
cant deal with genes that were not selected from sources other then
the annotation view. A view based approach to source appears to be
faulty. Need to somehow look at things from a model perspective, or
have to go digging into the view to see if the selection is present there.

This is why delete selection (and other editing functions) doesnt work
when the selection source is from the gene menu.