Fixed bug where external selection(from annot tree, gene menu...) would not get deleted. The problem was that Selection.getSelectedData(View) was being used. This gets the subpart of the Selection where the source is from that view, with the idea that it would get everything selected in that view. But external selections cause annots to be selected but not have annotView as its source and thus dont get picked up.
I changed this to make the retrieving of a subset of the selection
model based rather than view based (though there is more work to be
done here). So now annotation editor is given the model object at the
top/root of the model for that view, which for annotation view is a
GenericAnnotationSet which holds all the genes (how bout a GeneHolder
interface with a getGenes()?). Annot editor queries the selection set
with this and selection set looks for descendants of it in the
selection set. Added isDescendantOf to SeqFeatureI for this purpose.
Also changed contains to isAncestorOf as a different contains is
already in Range.