Coupled to a corresponding change in the controller-
Before: ApolloPanel receives a click, fires an event at controller, listens to
the corresponding event fired by the controller at it, and _then_ repaints.

Now: ApolloPanel receives a click, fires an event at controller, is _deaf_ to the
resulting FeatureSelectionEvent fired back by the Controller, and _then_ repaints
(after firing the event).

This just seems cleaner - a source should not have to receive the events it fires
in order to do stuff (like repaint or highlight feature edges).