The sequence window(SeqExport) now will show the transcript sequence if all the exons of a transcript are selected. In the past it showed all of the exon sequences. Bev and Sima were not too keen on this. What was funny about it is that if you selected the transcript (double click or intron) you got the transcript sequence, but if you rubberbanded or shift clicked the transcript you got exon sequence, even though both selections appear the same. This is now ammended. The analogous is also done for results.
Selection now has getConsolidatedFeatures, that adds a parent and
deletes kids if all the kids are present (except for holders). This
is not done for every selection, as it would make selection crawl
(particularly large rubberbands). SeqExport calls this method as a
post processing step, so the consolidation only happens if you are
showing the sequence window. I think at the moment the sequence window is the
only view that needs this consolidation, but its in Selection so any
other view can ask for the consolidated selection(might want to cache
if more than one views want it).