When selecting a transcript, the start and stop codons that line up with the start and stop of translation now highlight(if they are present of course). This was functionality that was present in the old "berkeley branch".
I decided to do this with a  handleFeatureSelection in SiteView as it
seemed simpler that way. Otherwise apollo panel would have to shoot
off the selection to siteviews, and make sure they got deselected.

I took out SelectViewI as an interface of SiteView. At the moment
there is no external way to select a codon. If this changes we can put
SelectViewI and its select back in.

So I mimiced the functionality I found in the old berkeley branch
which is that codon selection happens with exon or transcript selection.
So if you select and "zoom to select" on an exon it will actually zoom
to the exon and the associated codons of the transcript, so you will
pretty much zoom to the transcript not the exon. Is that problematic?

This is even funnier with the scroll/zoom bug. Codons lose their
selection on zooming and scrolling. So when you zoom to the exon and
its codons, the codons are no longer selected. This i will fix soon.

So to be consistent with how selection happens on mouse clicking and
to handle deselection, the selected codons from transcript selection
are packaged up as apollo.gui.drawable.SiteCodons and added to the
current Selection in SelectionManager (just like for mouse
clicking/findForSelection).

At first I just did a select() with this
new Selection. This caused big selections that involved transcripts
(or big selections that transcripts were added to) to be really
slow. The reason was select was firing a selection event, which was
causing apollo panel to unnecasarily refind everything in the
selection again.

So I added addToCurrentSelection to selectionManager which I used to
add the codons without firing a selection event. This assumes that
no other views care about codon selection which is true at the moment.

An alternative approach worth considering for non exclusive selection
is to only fire a selection event with whats new to the
selection. This is possible with selection events now carrying
exclusivity information(whether or not selection replaces or adds to
current selection).