Exons that were out of range (outside of curation set) had funny frame numbers, which are displayed in the annotation tree. I changed it to say "out of range" in this case.
Stumbled on an interesting issue. I used
getRefSeqSequence().getRange().contains(exon) to see if the exon is
out of range. This came back with all reverse strand features being
out of range. This was because the contains function in Range
(superclass of CurationSet) was checking if the strands were the same
and CurationSets strand was forward. So I overrode contains in CurationSet
as well as overlaps so it ignores strand. Basically this is saying
CurationSet is a special Range that is strandless, can have features
from both strands.