The different annot types(genes,transposons,trnas...) were previously all lumped together visually. Getting them differentiated is a bit tricky. At the moment all these different types are being shoved into the Gene/Transcript/Exon heirarchy for expediency. The game xml does not specify a type for the artificial transcript and exon of the non-gene types. Given no type Exon defaults to "exon" and Transcript "transcript".
To enable visual control of the alternate types in this kludgy state, the default types
for exon and transcript have been removed. If exon and transcript dont have a type
they query their parents type.

Also theres some commented out lines on the beginnings of an attempt to make alternate
types GenericAnnotationSets rather than Genes. We've decided to hold off on that for now,
but eventually we will probably need to go down that route to handle more complex types.