Added a zigzag "shape" for transposons. I did a little experimenting here which we can certainly take back if people dont like it. I added a shape directory to the drawable directory. In it is a shape interface and a zigzag class which implements the interface. For now the only thing in the interface is a draw method.
So the idea is this. Rather than having different drawables for every glyph/shape,
you plug lightwieght shpaes into the drawable. So the drawable handles all the
mechanics of being a drawable (like being a SeqFeatureI) and defers the drawing
to its shape.  This means you can plug different shapes into
all the drawables. At the moment DrawableSeqFeature checks to see if it has a shape and
and if it doesnt, it does its normal draw.

The zig zag shape itself could probably use a little more work, the lines could be
thicker and the line at the end could be extrapolated rather than going to the top.

So for now the tiers file says "shape:Zigzag", so DrawableFeatureSet looks for the
shape tag, and for now it just assumes its in apollo.gui.drawable.shape. These
are easily fixable.

Well just thought Id try this way to see if it could work.