Added in AlignSubMenu to TiersPopupMenu, which has 2 options: "Align the selection" or "Align the features of the same type in the selected region". The second option finds all features contained by the low and high of selection that have a visual type thats part of the selection.
SeqFeature.getType is not used for the gathering of features by types. The reason is
that SeqFeature.getType returns the "logical" type, not the "visual" type. The logical
type is the type from the data, the visual type is what apollo maps the logical type into
guided by the tiers file. In the tiers file the logical types are listed in the curly brackets
and the visual type that these visual types map to is the string before the curlies.

type "Fly Sequence", "DGC cDNA", {"sim4:na_DGC.dros" "sim4:na_cDNA.dros" "sim4:na_DGC-cDNA.dros"},...

in this line from tiers the sim4 strings are the logical types and "DGC cDNA" is the
visual type. I was tempted at 1st to add a getVisualType() to SeqFeatureI, but this would not be
right having visual information in the datamodel.
Config.getPropertyScheme().getFeatureProperty(feature.getType()).getDisplayType()
maps logical type to visual type. DetailInfo.getPropertyType(feat) is a convenience
function for this. The reason Im writing all this is it took me a while to figure this all out.

Need to add this menu to annot menu as well. Also it currently only takes in results, needs to
take in annots as well.

We can check out if we like the new submenu - nomi what do ya think?
Right now its slow with lots of features as its not chopping out introns which
micheles stuff will soon do.