Some game sequences were coming up with N's in jalview, Ive ammended this. getHitSequence is now called on the feature pair, which tries both getHitFeature().getFeatureSequence() (which is synonomous with getHitFeature().getRefSequence()) and if it doesnt find sequence there it gets it from the alignment sequence. This is because some game sequences awkwardly have alignments but no ref seq. So this should be fine with ensembl because it first makes the getRefSeq call that was previously in there.
The other change was that its just calling getResidues() instead of
getResidues(fp.getHstart(),fp.getHend()). The hstart and hend were not
in the same coordinates as the sequence it was getting from game. Is
this problematic? Is this a bug in game? Will this mess up ensembl? I
would test in ensembl if I could, but I dont have pserver access of
course. Steve and/or Michele could you try it out and make sure I
didnt mess things up for ensembl. If it is I'll rework things
but I think it should be ok.