Added isProtein as a tier property so one can tell apollo a tier is protein without having to stick "Protein" in the tier label. This gets away from the hard wiring of proteins. The problem came up for a user this morning and it was trivial to add so I did.
This is added to the tier as "key=value:isProtein=true". So Ive
introduced the idea of key-value pairs to the tiers. This gets away
from the rigid ordering of arbitrary properties and its clearer since
the key will say what the value is for. A tag was necasary for now
("key=value:") to distinguish it for parsing. New properties should
use key-value, and it'll be pretty easy to convert the old ones at
some point. This is a step towards making tiers more managable.