filtering of BLAST results now working on expect, score, percent_identity, and separating of HSPs into individual hits if so desired (for genome to genome analysis). few small changes to the gui (still needs history to be added).
also moved PropertyScheme out of Config and into Style. This way it
persists between opens. Before it was creating a new PropertyScheme
and reparsing the tiers.dat file every single time something new was
loaded. Needed to do this in preparation for layering of BLAST results
which does not want to replaced the current set of tiers and types,
but merely extend them. Now that the PropertyScheme is retained this
is possible.