Some peptide alignments were being seen as dna by jalview. The reason was it was testing for basepairs per residue assuming hit start and hit end were in the same coordinates as the alignment. In game there are tblastx features where the alignment is peptide and the hit sequence is nucleotide, so you cant use the hit start and end. Changed it to use length of alignment sequence, which should be universal. Added alignmentIsPeptide to FeaturePair. Added calculateResidueTypeFromFeatureLength to Sequence which calculates residue type from the length of the feature(if there is more 2 basepairs per residue its called a peptide). This formalizes what Michele called: // Nasty hack to tell peptide from dna. Blergh.
But if the data does not tell apollo what type of sequence it is then
what choice do we have. Ideally the data should tell us the
residueType, but currently I dont think any of the adapters do. So
this is the best we can do given the lack of info. If there are
sequences of type RNA that will definitely have to be put into the data.