for BLAST filtering added capapbility to throw away hits with low information content (bassed on humffman inencoding)
added the final (hopefully) changes need to submit the
hetermochromatic sequence to genbank. this involved one
change to the gameadapter and one to the gamesequence.
the gameadapter was creating the sequence model twice.
once when it was reading it in (as a generic Sequence model)
and again as a gameseqeunce. the problem was it
did not capture all of the attributes that had been
read in by the XML. much information was being lost
as  a result of the transfer. now it creates a gamesequence
from the start so that all attributes are captured. one
method needed to be added to the game sequence to allow
this for setResidues.