Encoded the Transcript kludge in DetailInfo.getPropertyType in FlyNameAdapter.getBioTypeForDisplay. I think name displaying issues belong in the name adapters, where a programmer can encode their own name preferences for their organism. So FeatureEditPanel now uses NameAdapter.getBioTypeForDisplay to get the type for its lower panel. In following the rules that seemed to be in place, if its a transcript of something other than a gene, than use the parents biotype. I tacked on "transcript" at the end, which seemed appropriate. Nomi feel free to adjust this to your liking.
Hope this makes the types issue easier. Why I feel it belongs in the
name adapter is that it seemed like display wanted it one with gene
transcripts and another way with all other transcripts. This didnt
seem like a general naming scheme to me, but something that was
appropriate for a name adapter.