Did a few things. I dealt with the issue where non fixed width fonts were not aligning properly. It turns out they would align properly if FastDraw was unselected. FastDraw is a much faster draw but can only be used by non fixed width fonts (eg Courier). So instead of the user having to know this, I automated it and took out the FastDraw menu option. So now the FontMetrics is queried for fixed width. I tried doing this by comparing '.' with getMaxAdvance but that didnt work(surprisingly). So I just compare '.' with 'A' which works for our purposes.
I added "slower" to Helvetica and TimesRoman and "faster" to Courier
to try to discourage people from using the slower fonts while still
giving them a choice. I tried to automate that but I couldnt get a
hold of a graphics and its font metrics at menu init time, so for now
its hard wired.

So I also went after the issue of getting the menu check boxes to act
like radio buttons, meaning when one is selected the others
deselect. Without this its hard to know what state things are in. Well
this turned out to be a bit of a humdinger. Its now fixed but it was a
bit convuluted. It probably could use a refactor but I dont have time
right now. Steve and Michele, I would definitely appreciate your
feedback on this, especially if theres already a way to do this which
I didnt see.

So in JalToggleGroup.stateChanged before it does a
setState(current,true) I put in a deselectAllItems.
deselectAllItems loops through all the items and deselects them. First
I tried setStateQueitly(false) which did not unset the checkbox.
Then I did setState(false) this caused an infinite loop. In
JalToggleAction.setState there is this:
  public void setState(boolean state) {
    setStateQuietly(state);
    if (group != null) {
      group.stateChanged(this);
    } else {
      fireJalActionEvent(new JalActionEvent(this,this,JalActionEvent.SELECT));
    }
  }

So it was just calling the group.stateChanged and looping. What I
wanted was the action event, which makes it to the check box. So I
scrapped the setState and did this:

      if (t instanceof JalAction) { // i think this is always true
	JalAction ja = (JalAction)t;
	ja.fireJalActionEvent(new JalActionEvent(ja,ja,JalActionEvent.SELECT));
      }

Which still caused an infinite loop. But I was getting somewhere. The
checkbox was getting set, but then it was alarming its listeners,
which would cause an infinite loop. So in JalCheckBoxMenuItem i
disabled the listener on a handleJalAction(JalActionEvent) and then
readd it after setting the state:

  public void handleJalAction(JalActionEvent evt) {
    //System.out.println("JalCheckBoxMenuItem got a handleJalAction");
    if (!action.isEnabled()) {
      disable();
    } else {
      enable();
    }
    if (action instanceof ToggleI) {
      if (((ToggleI)action).getState() != getState()) {
	removeItemListener((JalToggleAction)action); // otherwise infinite loop
        setState(((ToggleI)action).getState());
	addItemListener((JalToggleAction)action);
      }
    }
  }

This seems to me what has to happen to avoid the infinite loops. Some
of the things im doing could probably be fomralized with a real
method. So was there already a way set up to do this that Im not seeing?