A very quick fix for the following bug:
AE34) If you add a new exon to an existing transcript (using shift-drag),
it shows up abnormally thin, and in the color of the original result!
Also, when you click on the edited gene, the righthand detail panel shows
scores for the exons, which it doesn't do with unaltered genes.

The problem was our types kludge to get non gene types in there. This
is a quick kludgy fix for the release(a kludge on top of a kludge if
you will - gasp). Ill revisit this post release
as our types kludge is falling apart - I suspect the problem with the
backup file is also related to it.