A few things here: So I've changed how type and biotype work, and put them in line with suzis vision for them. The problem was the code was trying 2 different questions of the types. One was what are you a part of, and the other was where are we in the heirarchy. For example for a Transcript sometimes we wanted to know that it was a transcript other times we wanted to know what its context was, was it part of a gene, transposon, trna or what?
Suzi introduced biotype to deal with this issue. The problem was
biotype was giving all sorts of different answers, and was not giving
the responses it suppose to. The idea was that biotype would answer
the general type question. That for an exon its type would be exon but
its biotype would be thing its a part of, eg gene, transposon, etc. So
Ive changed it to reflect this. Type (queried with getType()) for exon
and transcript is always exon and transcript. BioType for exon and
transcript now query the gene object for its biotype. Type and biotype
in the Gene object are the same (eg gene, tRNA, transposon...). This
makes dealing with non gene types a lot easier and has ironed out some
bugs from our previous hack.

So in theory exon and transcript do not have to be in the tiers file
anymore, as they are covered by gene being in there. Ill check that
out.

I checked it out and this didnt break the other dataadapters. The
other datadapters i dont think set the biotype, which is ok because
biotype for exon and transcript take on genes biotype, and gene and
all other seq features take their biotype from the type. So if you
dont set the biotype at all it should work.

So this effects the colors of annots, and colors in the evidence
panel. It also effects the names in annot info and in the evidence
panel. I changed a method in the FlyNameAdapter. I think I have worked
out all the issues there. Nomi could you double check me on this.

The auto save serialized backup file now seems to work again with
these changes. This was broken with the previous types hack.

I think ive fixed the changing of types in annot info getting
reflected in the main window. The only kink left to fix is that the
SetDetailPanel doesnt refresh after an annot info commit.


In the GAME adapter I mapped the 3 level game xml for a miscelaneous
curator observation into just a GenericAnnotation, where it normally
maps into a Gene-Transcript-Exon structure. This exposed lots of
places in apollo that are hard wired for annotations to
Gene-Transcript-Exon. Ive ammended a lot of these but not all of
them.

Some functionality had to migrated to GenericAnnotation from
other classes (like GenericAnnotationSet).

FeatureEditorDialog (annotation info editor) was hardwired for
gene-transcript-exons, but I have it now working with
GenericAnnotations.

AnnotationMenu was hardwired for genes and transcripts for all of its
options. The Sequence window and the gene report now work with a
GenericAnnotation.

This is a work in progress. There is still work to be done here. Since
this is not fully working yet I put a flag in GAMEAdapter called
TRANSFORM_NON_GENE_TYPES which turns off the transformation. Ive
checked it in with false, turned off. When someone is working on this
obviously they should turn it on.

Still to be done...
This is only doing misc curator observation. All the other types that
are not Gene-Transcript-Exon have to be transformed, as the game xml
is not going to change for these types. If we get a chado adapter
presuambly this data transformation will also not need to happen.

They have to be transformed back for saving from the new apollo data
structure to that which is expected in the xml. This is true of misc
cur observation as well as ive only done the mapping in the load
direction no the save direction. Without doing it in the save
direction the xml will get saved with the new apollo data structure.

jalview, the EDE and editing all need to get working with the new
types.

Ive only experimented with a flat GenericAnnotation. Have not done
nested alternative types (eg a 4 or 5 level transposon). Also a
promoter hanging off a transcript needs to be tried out.