Fixed jalview reverse strand issues. It was a bit of work.  Genomic reverse strand translations are now correct. It is now displaying 5' to 3' per annotators request.
All the data structures in CondensedAlignment were
getting more confusing with reverse strand issues thrown in so I
created the Block class that tracks all data with a block (features,
genomic start/end, jalview start/end, strand), which made things
easier.

ResidueProperties had parts of the genetic code wrong, so I made
corrections.

Mapper works for both strands.

Alignments still not right with insertions and deletions - need
cigars.

There was a subtracting of one in getting genomic residues that seemed
to be wrong and creating problems so I took it out.

sortByStart is now strandedSort and sorts in descending direction for
reverse strand.