(somewhat) better way of checking whether we're about to run out of memory. Using freeMemory() turns out not a good way to check free memory, because the heap might still be at its initial smaller size.  We might have almost used up that initial heap, but then the JVM can allocate more memory, up to the max.
So instead I'm now assuming that the memory setting in apollo.cfg is in fact
the max memory heap allocated to Apollo, although that is not
absolutely guaranteed to be true, because it depends on how the memory was
set in the installer.  But I don't know any other way to ask the size of
the max heap.

The weird thing is that the memory used can end up *higher* than the supposedly
maximum heap size, so you can get a free memory value that's less than zero.
But that's still a bad sign, so hopefully our warnings will be more appropriate
than before.