So I set out to fix the jalview bug where all the sequences were getting off by one on scrolling and it turned into a slippery slope. So Suzi had changed jalviews Sequence (that was independent of apollos Sequence) to AlignSequence which subclassed from apollo.datamodel.Sequence. This is nice because it gets rid of redundant sequence models, but it introduced this off by one bug.
The problem was that jalview sequences are zero based and apollo
sequence are either one based or they have their own range, but they
dont deal with sequences that are zero based(with no range).

Theres two things that have to happen with sequence.
1) The end has to be made inclusive by adding one to it. This is
because apollos substring ends are inclusive where java Strings ends
are exclusive.
2) Before doing java string operations(ie substring), the coordinates have to be set
to zero based. If they are one based, one has to be subtracted from
start and end. If they are range based, the range start has to be
subtracted from start and end. If they are zero based(jalview) dont do anything.

I did a little refactoring as I found things a bit confusing how they
were and it took me a while to figure out what was going
on. AbstractSequence.pegLimits checked if the coordinate was in
bounds, but it also had the side effect of doing the shifting and end
inclusion. Previously AbstractSequence.getResidues(start,end) called
pegLimits(start,true). The true would cause pegLimits to subtract one
from the start if the sequence had no range. pegLimits on end was
called with false, so it would not get one subtracted. So thus it
would both be zero shifted and end included. But end inclusion was
only for non-range seqs, so seqs with range would still have to do
there own end inclusion, and you just have to know this.

I changed pegLimits to do none of the side effects of shifting, and it just checks
limits now. I added a makeEndInclusive method (which just adds one,
its there to make it really clear why its doing it). I added a
needInclusiveEnd method which by default is true. EnsCGISequence
overrides this to return false because the cgi takes care of inclusive end (no java
string issues). GAMESequence no longer adds one to its end as its now
done in AbstractSequence. CacheSeqIdLoader (which can only handle
GAMESequence and perhaps should be renamed CacheGameSeqLoader) was
adjusted to deal with the inclusive end.

I also added shiftFromOneToZeroBased(int) which just subtracts one as
well, again its to make explicit what is going on. And theres a
needShiftFromOneToZeroBasedCoords() and a
setNeedShiftFromOneToZeroBasedCoords(boolean). Seqs that have a range
return false for needShiftFromOneToZeroBasedCoords as they do the
shifting themselves. Rangeless seqs default to return true for
needShifting. Zero based seqs(AlignSeq) setNeedShift to false. While
this is clearer than before, now that im writing this im seeing it
could be done even clearer, maybe ill fiddle with it a bit more.

So now AbstractSequence can deal with zero-based seqs which fixes the
bug in jalview. Also another bug fixed is that result seqs were coming
up with an extra base pair. It turns out that inclusive end was being
compensated for twice, both in AbstractSequence and Sequence. So I
took that out of Sequence.

Steve,Michele, or Vivek: could you please test these changes on ensembl results
(and other data adapters). Either that or get result seqs visible from
the outside world - i know the pfetch server is up but its still not
installed in apollo to view the result seqs - steve do i have this
right? Ive tested it with the genomic from ensembl and thats working
ok.

I also noticed the very last base of the genomic sequence is not
getting drawn, but it will select and find and even draw if coloring
by splice site potential, so im going to look into that.