Fixed null pointer exception on EvidencePanel repaint after a Transcript delete. The evidence panel was still displaying the deleted transcript which threw a null pointer since it has a null gene, and it needs to get its biotype from its gene. So I added in some null checks to Transcript and DetailInfo. But I went a step further than that.
It seemed problematic to me that the deleted transcript was still
being displayed in the evidence panel. And if you clicked on it you
got exceptions (which has always been true, unlike the exception i
introduced above). So I changed evidence panel to clear out anything
deleted. I did this actually at the Selection level. The
SelectionManager now listens for annotationChangeEvents and if a
deleted item is in the selection, it removes it from the selection, and
sends out the modified selection in a selection event.

This means that it could be sending out an empty selection, which needs to mean clear
out the selection. Evidence panel was ignoring empty selections so i
changed that as well. But this has had the effect that if you click on
blank space which deselects anything selected, the evidence panel goes
blank. I think its better that way as its more accurately reflecting
the state of selection, but if people dont like it its easy enough to
change. (though i think moving the red sliders should not deselect
feats - which shouldnt be hard to change -  on the todo)