Lots of changes. All having to do with bop.
The biggest structural change was reorgnizing datamodel and util
so that NO gui specific code from apollo was resident any longer.
that way bop (and other applications) can be compiled without
bringing in a mountain of totally unrelated code, simply because
part of the datamodel or util package knows about the GUI. Two
things to look for:
1. naming peptides is appl. specific and should not be handled in
the transcript datamodel, but some where in the appl itself
2. the featurepair should not be doing a check to see if the sequence
is a SRS sequence before it decides how to do a toString method.
thats gross.

bop runs as a stand-alone appl. or as a part of apollo. more to come