Did some repackaging of synteny: apollo.gui -> apollo.gui.synteny CompositeApolloFrame(soon to be merged with ApolloFrame), SyntenyLayout, SyntenyLinkPanel, SyntenyPanel
gui -> dataadapter.synteny
ChromosomeDrawer, ChoromosomePanel, FullEnsJDBCSyntenyPanel
These are gui things but they are solely for the data adapter. Our
convention is to put dataadapter gui stuff in dataadapter not
apollo.gui. There could be a dataadapter.synteny.gui if we wanted to
go all the way with it, but doesnt seem too bad to have it in synteny.

removed dataadapter/CompositeAdapterUI - not used anymore
should probably remove FullSyntenyPanel as well
maybe rename SyntenyMenu ChoromosomePanelMenu?