It turns out that _sometimes_ the 'characters' method of a sax-parser can be called multiple times within the one tag. This prompts the following broad change:
The handleCharacters() method in all these TagHandlers has been moved
to the TagHandler superclass: this class 'accumulates' the characters that are
sent to it via potential multiple invocations of 'handleCharacters' into
a cache. The tag then 'picks up' the accumulated character set when it does
handleEndElement() with a getCharacters() call.

Testing with read/retrieve shows no harm so far...