Suzi changed the background color of the loader.  I changed the title color because the default one wasn't showing up on her new background color.
I'm actually not convinced that it's a good idea to change the background color
of the loader, because the new color (a sort of muted turquoise) shows up in
some places, but other places still look gray, which looks kind of weird.  So
I'd be inclined to just leave it gray.