Fixed bug: - Start/stop codons can't be dropped on transcripts (works fine with JDK1.3).  This is true even on Macs. DrawableFeatureSets bounding box had a height of 0, due to getSize() returning 0. A rectangle with height 0 will not intersect in jdk 1.4, where it did in 1.3, thus why it broke in 1.4.
DrawableFeatureSet.getSize now returns the size of it largest
child. This fixed the drag and drop problem, as the rectangle now
intersected. But it introduced a new problem. The transcript/feature set lines
were then being drawn at the top of its children rather than in the
middle. Turns out DrawableFeatureSet and DrawableResultFeatureSet were
just using the y value of its bounding box. The reason this appeared
in the center was that the height of the rectangle was 0. So the
drawable feature sets are now querying for the center of the tier
with:

transformer.toPixelY(getYCentre(manager))

Cant use middle of rectangle as different feat sets are different
heights, so when they collapse the transcript line will not all
collapse into one.

One thing I noticed in the middle of this is that when dashed feature
set lines collapse they collpase into a solid line. If the dashes were
in synch they would collapse as a dashed line.